src/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.35.7
networkx==3.4.2
python-dotenv==1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
slowapi==0.1.9
httpx==0.27.0
